+++ src/__generated__/CollectionHeader_collection.graphql.ts
@@ -1,0 +1,70 @@
+/* tslint:disable */
+
+import { ReaderFragment } from "relay-runtime";
+declare const _CollectionHeader_collection$ref: unique symbol;
+export type CollectionHeader_collection$ref = typeof _CollectionHeader_collection$ref;
+export type CollectionHeader_collection = {
+    readonly slug: string;
+    readonly title: string;
+    readonly description: string | null;
+    readonly headerImage: string | null;
+    readonly category: string;
+    readonly credit: string | null;
+    readonly " $refType": CollectionHeader_collection$ref;
+};
+
+
+
+const node: ReaderFragment = {
+  "kind": "Fragment",
+  "name": "CollectionHeader_collection",
+  "type": "MarketingCollection",
+  "metadata": null,
+  "argumentDefinitions": [],
+  "selections": [
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "slug",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "title",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "description",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "headerImage",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "category",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "credit",
+      "args": null,
+      "storageKey": null
+    }
+  ]
+};
+(node as any).hash = '9b4f64d7222b3912ff2f6c9d1b569a4d';
+export default node;