+++ src/__generated__/ShowTestsQuery.graphql.ts
@@ -151,18 +151,18 @@ fragment ShowArtistsPreview_show on Show {
   internalID
   slug
   artists {
+    id
     internalID
     slug
     href
     ...ArtistListItem_artist
-    id
   }
   artists_without_artworks: artistsWithoutArtworks {
+    id
     internalID
     slug
     href
     ...ArtistListItem_artist
-    id
   }
 }
 
@@ -882,10 +882,10 @@ return {
             "concreteType": "Artist",
             "plural": true,
             "selections": [
+              (v7/*: any*/),
               (v1/*: any*/),
               (v2/*: any*/),
               (v12/*: any*/),
-              (v7/*: any*/),
               (v3/*: any*/),
               (v13/*: any*/),
               (v8/*: any*/),
@@ -1135,7 +1135,7 @@ return {
   "params": {
     "operationKind": "query",
     "name": "ShowTestsQuery",
-    "id": "afc5c42c803ac48cc2463ab20bb8bd16",
+    "id": "a53f4aae0b0be77d7be03e2a04353567",
     "text": null,
     "metadata": {}
   }