+++ src/lib/Scenes/Collection/index.tsx
@@ -1,0 +1,1 @@
+export { CollectionContainer } from "./Collection"